python: move flux-mini functional code to flux.cli
Problem: The code for all job submission commands is stuffed into a
single source file in src/cmd/flux-mini.py. This makes the file large
and difficult to maintain, and makes it impossible to split the one
large `flux mini` command into smaller commands at the top level of
the `flux` command hiearachy.

Split `flux-mini.py` into several modules under `flux.cli`. Place
the base classes used by several of the subcommands into a base.py
module, and locate other subcommands in their own files so the code
is more organized.

Since all these new classes will be eventually used outside of
flux-mini, update any documentation or output strings that mention
flux-mini to be more generalized.

Update affected tests in the testsuite.

Refactor `flux-mini.py` so it uses the new flux.cli modules.
